# go-api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a main payment-gateway API .",
        "title": "Payment Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/accounts": {
            "get": {
                "description": "Returns a list of all accounts, with pagination support.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get all accounts with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Account"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new account with a username.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Create a new account",
                "parameters": [
                    {
                        "description": "Account data for creation",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    }
                }
            }
        },
        "/cards": {
            "post": {
                "description": "Creates a new, fictional card and associates it with an account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Create a new card",
                "parameters": [
                    {
                        "description": "Account ID to associate the new card",
                        "name": "card",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Card"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to create card",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        },
        "/cards/{accountId}": {
            "get": {
                "description": "Returns a list of all cards associated with an account, ordered by creation date.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Get all cards by account ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the account",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Card"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation failed",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve cards",
                        "schema": {
                            "$ref": "#/definitions/api.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.CreateAccountRequest": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "username": {
                    "description": "@Description The username of the new account.\n@Example charlie",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
        "dto.CreateCardRequest": {
            "type": "object",
            "required": [
                "account_id"
            ],
            "properties": {
                "account_id": {
                    "description": "@Description Account ID to associate the new card\n@Example e252f5dd-ded2-4a30-a4a5-6e2940008d54",
                    "type": "string"
                }
            }
        },
        "models.Account": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "@Description The creation timestamp of the account.\n@Example 2025-09-22T19:15:24.526505Z",
                    "type": "string"
                },
                "id": {
                    "description": "@Description The unique identifier of the account.\n@Example 550e8400-e29b-41d4-a716-446655440000",
                    "type": "string"
                },
                "updated_at": {
                    "description": "@Description The last update timestamp of the account.\n@Example 2025-09-22T19:15:24.526505Z",
                    "type": "string"
                },
                "username": {
                    "description": "@Description The unique username for the account.\n@Example charlie",
                    "type": "string"
                }
            }
        },
        "models.Card": {
            "type": "object",
            "properties": {
                "account_id": {
                    "description": "@Description The unique identifier of the account this card belongs to.\n@Example 550e8400-e29b-41d4-a716-446655440000",
                    "type": "string"
                },
                "card_token": {
                    "description": "@Description The unique, irreversible token for the card.\n@Example 5b7c16af7278094cd14bd041079111ed00fa832c8a460d8e3f40156408d99475",
                    "type": "string"
                },
                "created_at": {
                    "description": "@Description The creation timestamp of the card.\n@Example 2025-09-22T19:15:24.526505Z",
                    "type": "string"
                },
                "id": {
                    "description": "@Description The unique identifier of the card.\n@Example 550e8400-e29b-41d4-a716-446655440000",
                    "type": "string"
                },
                "last_four_digits": {
                    "description": "@Description The last four digits of the card number for identification.\n@Example 8995",
                    "type": "string"
                }
            }
        }
    }
}